Add more missing Exceptions, moved some exceptions to a folder alligned with the use case for them
