--- backend/src/main/java/com/servicedesk/lite/org/exception/UserNotFoundException.java
@@ -1,7 +1,0 @@
-package com.servicedesk.lite.org.exception;
-
-public class UserNotFoundException extends RuntimeException {
-    public UserNotFoundException(String message) {
-        super(message);
-    }
-}
